pyproject.toml
[project]
name = "z-quota"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.24.0",
    "aiohttp>=3.13.3",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
]
